fetch year to date chart from healthkit

+++ FitnessApp/Charts/Model/ChartsModel.swift
@@ -11,7 +11,7 @@ import Foundation
 struct DailyStepModel: Identifiable {
     let id = UUID()
     let date: Date
-    let count: Double
+    let count: Int
 }
 
 enum ChartOptions: String, CaseIterable {
